+++ lib/screens/editor/map_chooser.dart
@@ -1,7 +1,9 @@
 import 'dart:async';
 
 import 'package:every_door/constants.dart';
+import 'package:every_door/helpers/good_tags.dart';
 import 'package:every_door/models/amenity.dart';
+import 'package:every_door/providers/editor_mode.dart';
 import 'package:every_door/providers/geolocation.dart';
 import 'package:every_door/providers/imagery.dart';
 import 'package:every_door/providers/osm_data.dart';
@@ -60,10 +62,25 @@ class _MapChooserPageState extends ConsumerState<MapChooserPage> {
   updateNearest() async {
     final provider = ref.read(osmDataProvider);
     final filter = ref.read(poiFilterProvider);
+    final editorMode = ref.read(editorModeProvider);
     final location = center;
     // Query for amenities around the location.
     List<OsmChange> data =
         await provider.getElements(location, kVisibilityRadius);
+    // Filter for amenities (or not amenities).
+    data = data.where((e) {
+      switch (e.kind) {
+        case ElementKind.amenity:
+          return editorMode == EditorMode.poi;
+        case ElementKind.micro:
+          return editorMode == EditorMode.micromapping;
+        case ElementKind.building:
+        case ElementKind.entrance:
+          return false;
+        default:
+          return e.isNew;
+      }
+    }).toList();
     // Apply the building filter.
     if (filter.isNotEmpty) {
       data = data.where((e) => filter.matches(e)).toList();
@@ -137,11 +154,7 @@ class _MapChooserPageState extends ConsumerState<MapChooserPage> {
                   CircleMarker(
                     point: poi.location,
                     radius: 3.0,
-                    color: poi.isModified
-                        ? Colors.greenAccent
-                        : (poi.element?.isAmenity ?? true
-                            ? Colors.black
-                            : Colors.yellow),
+                    color: poi.isModified ? Colors.greenAccent : Colors.yellow,
                   ),
               ],
             ),